Added another controller

+++ api/Models/Customer.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace api.Models
 {
     public class Customer
@@ -6,7 +8,7 @@ namespace api.Models
         public string cfName {get; set;}
         public string clName {get; set;}
         public string cBusinessName {get; set;}
-        public int cPhoneNumber {get; set;}
+        public Int64 cPhoneNumber {get; set;}
         public string cEmail {get; set;}
         public string cUsername {get; set;}
         public string cPassword {get; set;}
